#20 Adding symbol and requires approval to leave type and use typeId and policyId in activatedType
